+++ enemies/RunningZombie.cs
@@ -193,11 +193,11 @@ public class RunningZombie : KinematicMonster
 		throw new NotImplementedException();
 	}	
 
-	protected override void OnDamage(Node2D node=null, int amount=0)
+	protected override void OnDamage(Node2D node=null,float amount=0f)
 	{
 		if(state!=STATE.damage&&state!=STATE.die)
 		{
-			base.OnDamage(node, amount);
+			base.OnDamage(node,amount);
 			animationDirection = Mathf.Sign(GlobalPosition.x - node.GlobalPosition.x);
 			animationPlayer.Play("HIT");
 		}
